Macro topo::root

macro_rules! root {
    ($($input:tt)*) => { ... };
}

Roots a topology at a particular callsite while calling the provided expression with the same convention as [call].

Normally, when a topological function is repeatedly invoked at the same callsite in a loop, each invocation receives a different [Id], as these invocations model siblings in the topology. The overall goal of this crate, however, is to provide imperative codepaths with stable identifiers across executions at the same callsite. In practice, we must have a root to the subtopology we are maintaining across these impure calls, and after each execution of the subtopology it must reset the state at its [Id] so that the next execution of the root is invoked at the same point under its parent as its previous execution was.

In this first example, a scope containing the loop can observe each separate loop iteration mutating count and the root closure mutating exit. The variables root_ids and child_ids observe the identifiers of the

struct LoopCount(usize);

let mut count = 0;
let mut exit = false;
let mut root_ids = HashSet::new();
let mut child_ids = HashMap::new();
while !exit {
    count += 1;
    topo::root!({
        root_ids.insert(topo::Id::current());
        assert_eq!(
            root_ids.len(),
            1,
            "the Id of this scope should be repeated, not incremented"
        );

        let outer_count = topo::Env::get::<LoopCount>().unwrap().0;
        assert!(outer_count <= 10);
        if outer_count == 10 {
            exit = true;
        }

        for i in 0..10 {
            topo::call!({
                let current_id = topo::Id::current();
                if outer_count > 1 {
                    assert_eq!(child_ids[&i], current_id);
                }
                child_ids.insert(i, current_id);
                assert!(
                    child_ids.len() <= 10,
                    "only 10 children should be observed across all loop iterations",
                );
            });
        }
        assert_eq!(child_ids.len(), 10);
    }, env! {
         LoopCount => LoopCount(count),
    });
    assert_eq!(child_ids.len(), 10);
    assert_eq!(root_ids.len(), 1);
}