Crate topo

topo creates a hierarchy of scoped, nested [environments][topo::Env] whose shape matches the function callgraph. These environments store singletons indexed by their type, and references to environmental values are available only to an enclosed call scope. When a #![topo::nested] function is called, its parent environment is cheaply propagated along with any additional values added at appropriate callsites.

Each environment in this hierarchy has a unique and deterministic [topo::Id] describing that environment and the path taken to arrive at its stack frame. These identifiers are derived from the path taken through the callgraph to the current location, and are stable across repeated invocations of the same execution paths.

By running the same topologically-nested functions in a loop, we can observe changes to the structure over time. The moxie crate uses these identifiers and environments to create persistent trees for rendering human interfaces.

Making functions nested within the call topology

Defining a topological function results in a macro definition for binding the function to each callsite where it is invoked. Define a topologically-nested function with the topo::nested attribute:

#[topo::nested]
fn basic_topo() -> topo::Id { topo::Id::current() }

#[topo::nested]
fn tier_two() -> topo::Id { basic_topo!() }

// each of these functions will be run in separately identified
// contexts as the source locations for their calls are different
let first = basic_topo!();
let second = basic_topo!();
assert_ne!(first, second);

let third = tier_two!();
let fourth = tier_two!();
assert_ne!(third, fourth);
assert_ne!(first, third);
assert_ne!(first, fourth);
assert_ne!(second, fourth);

Because topological functions must be sensitive to the location at which they're invoked and within their immediate parent, we transform the function definition into a macro to track the source location at which it is called. Future language features may make it possible to call topo-nested functions without any special syntax.

Requiring references from the environment

The from_env macro provides an attribute for functions that require access to a singleton in their environment. Here, the contrived function requires a u8 to add one to:

#[topo::from_env(num: &u8)]
fn env_num_plus_one() -> u8 {
    num + 1
}

topo::Env::add(1u8);
assert_eq!(env_num_plus_one(), 2u8);

This provides convenient sugar for values stored in the current Env as an alternative to thread-locals or a manually propagated context object. However this approach incurs a significant cost in that the following code will panic without the right type having been added to the environment:

// thread 'main' panicked at 'expected a value from the environment, found none'
env_num_plus_one();

See the attribute's documentation for more details, and please consider whether this is appropriate for your use case before taking it on as a dependency.

Macros

call
Calls the provided expression with an [Id] specific to the callsite, optionally passing additional environment values to the child scope.

callsite
Returns a value unique to the point of its invocation.

env
Declare additional environment values to expose to a child topological function's call tree.

root
Roots a topology at a particular callsite while calling the provided expression with the same convention as [call].

Structs

Callsite
A value unique to the source location where it is created.

Env
Immutable environment container for the current (sub)topology. Environment values can be provided by parent topological invocations (currently just with [call] and [root]), but child functions can only mutate their environment through interior mutability.

Id
Identifies an activation record in the current call topology.

Point
The root of a sub-graph within the overall topology formed at runtime by the call-graph of topological functions.

Functions

current_callsite_count
Returns the number of times this callsite has been seen as a child of the current Point.

Attribute Macros

from_env
Defines required [topo::Env] values for a function. Binds the provided types as if references to them were implicit function arguments.

nested
Transforms a function declaration into a topologically-nested function invoked with macro syntax to attach its call tree's (sub)topology to the parent topology.